requirements.txt
bcrypt==4.0.0
blinker==1.5
boto3==1.22.11
botocore==1.25.11
cachetools==5.2.1
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
cryptography==38.0.1
cssselect==1.2.0
cssutils==2.6.0
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.0
lxml==4.9.2
MarkupSafe==2.1.1
Pillow==9.1.0
premailer==3.10.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.5.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
SQLAlchemy==1.4.36
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.9
Werkzeug==2.1.2
WTForms==3.0.1
yagmail==0.15.293
zipp==3.8.0
